init defaults in tuple

+++ source/blender/blenkernel/intern/fcurve.c
@@ -1623,7 +1623,7 @@ static float dvar_eval_function(ChannelDriver *UNUSED(driver), DriverVar *dvar)
 
 	FnTuple fn_in = FN_tuple_for_input(fn);
 	FnTuple fn_out = FN_tuple_for_output(fn);
-	FN_tuple_set_int32(fn_in, 0, (int32_t)dvar);
+	FN_tuple_set_int32(fn_in, 0, (int64_t)dvar);
 	FnTupleCallBody callable = FN_function_get_callable(fn);
 
 	FN_function_call(callable, fn_in, fn_out);
